+++ src/main/java/net/runelite/client/plugins/microbot/microbotdashboardplus/window/DashboardWindow.java
@@ -396,7 +396,7 @@ public class DashboardWindow extends JFrame {
         footer.setBackground(ColorScheme.DARKER_GRAY_COLOR);
         footer.setBorder(new EmptyBorder(4, 10, 4, 10));
 
-        JLabel info = new JLabel("MicrobotDashboardPlus v0.3.1 - in-process poller, no HTTP");
+        JLabel info = new JLabel("MicrobotDashboardPlus v0.3.2 - in-process poller, no HTTP");
         info.setForeground(Color.GRAY);
         info.setFont(FontManager.getRunescapeSmallFont());
         footer.add(info);